% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csf.R
\name{csf}
\alias{csf}
\title{Generate output for a Chi Square Test for Goodness of Fit}
\usage{
csf(x, y = NULL, probs = NULL)
}
\arguments{
\item{x}{A formula ~Categorical Variable .}

\item{y}{A data frame, necessary if data frame is not referenced in formula or piped in.}

\item{probs}{A vector to test goodness of fit, must have same number of categories and add to 1 if percentages}
}
\value{
A list of output for reporting $analysis_type, $results, $fit_model, $observed, $expected, $standardized_residuals, $contribution, $table_percentages
}
\description{
This function performs a Chi Sqare test for Goodness of Fit on one factor and returns reporting information on the model.
}
\examples{
x2_results <- tea_tab \%>\% csf(~cool_time)
x2_results <- tea_tab \%>\% csf(~cool_time,probs = c(20,50,15,15))
x2_results <- tea_tab \%>\% csf(~cool_time,probs = c(.20,.50,.15,.15))
x2_results <- csf(tea_tab$cool_time)
x2_results <- csf(tea_tab$cool_time,probs = c(20,50,15,15))
}
