% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi.R
\name{csi}
\alias{csi}
\title{Generate output for a Chi Square Test of Association}
\usage{
csi(x, y = NULL, z = NULL)
}
\arguments{
\item{x}{A formula dependent variable ~ independent variable.}

\item{y}{A data frame, necessary if data frame is not referenced in formula or piped in.}

\item{z}{A data frame, necessary if data frame is not referenced in formula or piped in to support form var,var,df.}
}
\value{
A list of output for reporting $analysis_type, $results, $odds_ratio, $observed, $expected, $standardized_residuals, $contribution, $table_percentages, $column_percentages, $row_percentages
}
\description{
This function performs a Chi Sqare test for association on two factors and returns reporting information on the model.
}
\examples{
x2_results <- tea_tab \%>\% csi(cool_time ~ tea_type)
x2_results <- csi(cool_time ~ tea_type, tea_tab)
x2_results <- csi(tea_tab$cool_time, tea_tab$tea_type)
}
