% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_elapsed.R
\name{final_elapsed}
\alias{final_elapsed}
\title{This is a helper function for the final exam}
\usage{
final_elapsed()
}
\description{
This function will tell how much time has passed since loading the file
}
