% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cv.R
\name{format_cv}
\alias{format_cv}
\title{Format Cramer's V}
\usage{
format_cv(v, digits = 3)
}
\arguments{
\item{v}{Numeric value of Cramer's V.}

\item{digits}{Integer specifying the number of digits to display.}
}
\value{
A character string containing the formatted effect size.
}
\description{
Helper function used to format Cramer's V effect size values for
reporting. Values smaller than 0.001 are reported as \code{"V < .001"},
otherwise the value is rounded to the specified number of digits.
}
\keyword{internal}
