% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p.R
\name{format_p}
\alias{format_p}
\title{Format p-values}
\usage{
format_p(p, digits = 3)
}
\arguments{
\item{p}{Numeric p-value.}

\item{digits}{Integer specifying the number of digits to display.}
}
\value{
A character string containing the formatted p-value.
}
\description{
Helper function used to format p-values for reporting. Values smaller
than 0.001 are reported as \code{"p < .001"}, otherwise the value is rounded
to the specified number of digits.
}
\keyword{internal}
