% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idt.R
\name{idt}
\alias{idt}
\title{Generate Output for an Independent t-test}
\usage{
idt(x, y = NULL, tails = 2)
}
\arguments{
\item{x}{A formula dependent variable ~ dependent variable, a data frame variable can also be passed in.}

\item{y}{A data frame, necessary if data frame is not referenced in formula or piped in, if a data frame variable is passed in to x, a data frame variable will also be accepted.}

\item{tails}{An integer indicating whether the test is one-tailed '1' or two-tailed '2'}
}
\value{
A list of output for reporting $analysis_type, $results, $descriptive_statistics
}
\description{
This function will generate the output for an independent t-test.
}
\examples{
idt_results <- t_tall \%>\% idt(scones ~ tea)
idt_results <- idt(t_tall$scones ~ t_tall$tea, tails = 1)
idt_results <- idt(scones ~ tea, t_tall)
idt_results <- idt(t_tall, scones ~ tea)
}
