% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idw.R
\name{idw}
\alias{idw}
\title{Generate Output for an Wilcoxon Rank-Sum Test}
\usage{
idw(x, y = NULL, tails = 2)
}
\arguments{
\item{x}{A formula dependent variable ~ dependent variable, a data frame variable can also be passed in.}

\item{y}{A data frame, necessary if data frame is not referenced in formula or piped in, if a data frame variable is passed in to x, a data frame variable will also be accepted.}

\item{tails}{An integer indicating whether the test is one-tailed '1' or two-tailed '2'}
}
\value{
A list of output for reporting $analysis_type, $results, $descriptive_statistics
}
\description{
This function will generate the output for an Wilcoxon rank-sum test for independent non-parametric samples
}
\examples{
idw_results <- t_tall \%>\% idw(scones ~ tea)
idw_results <- idw(scones ~ tea, t_tall, tails = 1)
idw_results <- idw(t_tall, scones ~ tea)
idw_results <- idw(t_tall$scones ~ t_tall$tea)
}
