% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ost.R
\name{ost}
\alias{ost}
\title{Generate Output for a One Sample t-Test}
\usage{
ost(x, y = NULL, tails = 2, mu = 0)
}
\arguments{
\item{x}{a formula ~Dependent Variable (or DV~1) or a data frame column containing the Dependent Variable}

\item{y}{optional, a data frame when data frame is not referenced and a formula is passed to x or piped in}

\item{tails}{an integer indicating whether the test is one-tailed '1' or two-tailed '2'}

\item{mu}{a double representing the value you want to comare to, the default for this parameter is 0.0}
}
\value{
A list of output for reporting $analysis_type, $results, $descriptive_statistics
}
\description{
This function will generate the output for a one sample t test.
}
\examples{
ost_results <- t_tall \%>\% ost(~scones,mu = 44)
ost_results <- ost(t_tall$scones, tails = 1, mu = 44)
ost_results <- ost(t_tall$scones ~ 1, tails = 1, mu = 44)
ost_results <- ost(scones ~ 1, t_tall, tails = 1, mu = 44)
ost_results <- ost(~scones, t_tall, mu = 44)
ost_results <- ost(t_tall, ~scones, mu = 44)
ost_results <- ost(t_tall$scones, mu = 44)
}
