% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psw.R
\name{psw}
\alias{psw}
\title{Generate Output for Wilcoxon Signed-Rank Test}
\usage{
psw(x, y = NULL, z = NULL, tails = 2)
}
\arguments{
\item{x}{A data frame column with the first variable to compare, a formula can also be passed in (~Var or Var~1)}

\item{y}{A data frame column with the second variable to compare (~Var or Var~1), if a formula is passed in a data frame will be accepted}

\item{z}{A data frame if formulas are used in x and y and data is not piped in}

\item{tails}{an integer indicating whether the test is one-tailed '1' or two-tailed '2'}
}
\value{
A list of output for reporting $analysis_type, $results, $descriptive_statistics
}
\description{
This function will generate the output for a Wilcoxon Signed-Rank Test for non-parametric paired samples
}
\examples{
psw_results <- t_wide \%>\% psw(~black_tea,~green_tea)
psw_results <- psw(t_wide$black_tea, t_wide$green_tea, tails = 1)
psw_results <- psw(t_wide, ~black_tea, ~green_tea)
psw_results <- psw(~black_tea, ~green_tea, t_wide)
psw_results <- psw(t_wide$black_tea, t_wide$green_tea)
psw_results <- psw(black_tea ~ green_tea, t_wide)
}
