% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_backup.R
\name{rmd_backup}
\alias{rmd_backup}
\title{Backup all .Rmd files matching a pattern, if there is an exact copy already existing, with same hash, replace the most recent backup with this one timestamp is added as a suffix to the file name.}
\usage{
rmd_backup(search = "Group*")
}
\arguments{
\item{search}{File name to search for default is Group*, for a specific file only pass in file name with no file type e.g. Group_Analysis_Plan}
}
\description{
Backup all .Rmd files matching a pattern, if there is an exact copy already existing, with same hash, replace the most recent backup with this one timestamp is added as a suffix to the file name.
}
