% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{slr}
\alias{slr}
\title{Generate output for a Simple Linear Regression}
\usage{
slr(x, y = NULL)
}
\arguments{
\item{x}{A formula dependent variable ~ independent variable.}

\item{y}{A data frame, necessary if data frame is not referenced in formula or piped in.}
}
\value{
A list of output for reporting $analysis_type, $results, $linear_regression_model, $predictors, $coefficients
}
\description{
This function performs a simple linear regression and returns reporting information on the model.
}
\examples{
slr_results <- napData \%>\% slr(naptime ~ timestamp)
slr_results <- slr(naptime ~ timestamp, napData)
slr_results <- slr(napData, naptime ~ timestamp)
slr_results <- slr(napData$naptime ~ napData$timestamp)
}
