% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Return a list of 5 random words given a 4 digit numeric id}
\usage{
validate(ID4)
}
\arguments{
\item{ID4}{a 4 digit id this must be a number that is 4 digits long}
}
\value{
A string containing 5 random words and the ID that was passed into the function
}
\description{
This function is used during setup to validate that a student was able to install the eco230r package
}
\examples{
validation <- validate(1234)
}
