% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi.R
\name{csi}
\alias{csi}
\title{Generate output for a Chi Square Test of Association}
\usage{
csi(x, y = NULL, z = NULL)
}
\arguments{
\item{x}{A formula dependent variable ~ independent variable.}

\item{y}{A data frame, necessary if data frame is not referenced in formula or piped in.}
}
\value{
A list of output for reporting $analysis_type, $results, $odds_ratio, $observed, $expected, $standardized_residuals, $contribution, $table_percentages, $column_percentages, $row_percentages
}
\description{
This function performs a Chi Sqare test for association on two factors and returns reporting information on the model.
}
\examples{
x2_results <- x2_data \%>\% csi(class_rank ~ gender)
x2_results <- csi(class_rank ~ gender, x2_data)
x2_results <- csi(x2_data$class_rank, x2_data$gender)
}

