% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst.R
\name{pst}
\alias{pst}
\title{Generate Output for Paired Samples t-test}
\usage{
pst(x, y = NULL, tails = 2)
}
\arguments{
\item{x}{A data frame column with the first variable to compare, a formula can also be passed in}

\item{y}{A data frame column with the second variable to compare, if a formula is passed in a data frame will be accepted}

\item{tails}{an integer indicating whether the test is one-tailed '1' or two-tailed '2'}
}
\value{
A list of output for reporting $analysis_type, $results, $descriptive_statistics
}
\description{
'This function will generate the output for a one sample t test.
}
\examples{
pst_results <- pst(pst_data$black_tea, pst_data$green_tea, tails = 1)
pst_results <- pst(pst_data$black_tea, pst_data$green_tea, tails = 2)
pst_results <- pst(scones ~ tea, pst_data_tall, tails = 1)
}

