% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{slr}
\alias{slr}
\title{Generate output for a Simple Linear Regression}
\usage{
slr(formula, data = NULL)
}
\arguments{
\item{formula}{A formula dependent varialbe ~ independent variable.}

\item{data}{A data frame, necessary if data frame is not referenced in formula.}
}
\value{
A list of output for reporting $analysis_type, $results, $linear_regression_model, $predictors, $coefficients
}
\description{
This function performs a simple linear regression and returns reporting information on the model.
}
\examples{
slr_results <- slr(sales ~ adverts, sales)
slr_results < slr(sales$sales ~ sales$adverts)
}

